requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
amplitude-analytics==1.2.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
apprise==1.9.9
argcomplete==3.6.3
asgi-lifespan==2.1.0
asgiref==3.11.1
asyncpg==0.31.0
attr==0.3.1
attrs==26.1.0
azure-core==1.39.0
azure-storage-blob==12.28.0
beartype==0.22.9
black==26.3.1
bleach==5.0.1
boto3==1.42.95
botocore==1.42.95
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudpickle==3.1.2
colorama==0.4.6
convmerge==0.3.3
coolname==4.2.0
croniter==6.2.2
cronsim==2.7
cryptography==46.0.7
cyclopts==4.11.0
datamodel-code-generator==0.26.1
datasets==4.8.4
dateparser==1.4.0
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
Django==5.1.15
django-annoying==0.10.6
django-cors-headers==4.7.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-ranged-fileresponse==0.1.2
django-rq==3.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.18.0
docutils==0.22.4
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-spectacular==0.28.0
email-validator==2.3.0
exceptiongroup==1.3.1
expiringdict==1.2.2
Faker==40.15.0
fakeredis==2.35.1
fastapi==0.136.1
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.2.0
genson==1.3.0
google-api-core==2.30.3
google-auth==2.49.2
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-core==2.5.1
google-cloud-logging==3.15.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.4.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
humanize==4.15.0
hyperframe==6.1.0
idna==3.13
ijson==3.5.0
importlib_metadata==8.7.1
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
jsf==0.11.2
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
# label-studio server pins conflict with prefect/pydocket (python-json-logger) and dateparser (pytz).
# Run the LS app via Docker or a separate venv; keep the SDK for API/export tooling.
label-studio-sdk==2.0.18
launchdarkly-server-sdk==8.2.1
lupa==2.8
lxml==6.1.0
lxml_html_clean==0.4.4
Mako==1.3.11
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
openai==1.109.1
opencv-python-headless==4.13.0.92
opentelemetry-api==1.41.0
ordered-set==4.0.2
orjson==3.11.8
packaging==26.0
pandas==3.0.2
pathspec==1.1.0
pendulum==3.2.0
perfect==0.0.1
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
prefect==3.6.27
prometheus_client==0.25.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psycopg==3.3.3
psycopg-binary==3.3.3
py-key-value-aio==0.4.4
pyarrow==22.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyboxen==1.3.0
pycparser==3.0
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
pydocket==0.19.2
Pygments==2.20.0
PyJWT==2.12.1
pyRFC3339==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
python-slugify==8.0.4
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
readchar==4.2.2
redis==5.2.1
referencing==0.37.0
regex==2026.4.4
requests==2.32.5
requests-file==3.0.1
requests-mock==1.12.1
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.4
rich-rst==1.3.2
rpds-py==0.30.0
rq==2.6.1
rstr==3.2.2
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
rules==3.4
s3transfer==0.16.1
semver==3.0.4
sentry-sdk==2.58.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
text-unidecode==1.3
tldextract==5.3.1
toml==0.10.2
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ua-parser==1.0.2
ua-parser-builtins==202603
ujson==5.12.0
uncalled-for==0.3.1
uritemplate==4.2.0
urllib3==2.6.3
user-agents==2.2.0
uuid_utils==0.14.1
uvicorn==0.46.0
webencodings==0.5.1
websockets==16.0
wheel==0.47.0
wrapt==2.1.2
xmljson==0.2.1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
